move open in explore to its own file
